Make `Transformation`s classes chainable to form `TransformationList`s
This removes some boilerplate in construction of singleton
transformation lists for the purpose of further composition.
